Fix the engagement tests
