+++ app-modules/engagement/tests/Tenant/EngagementFile/RequestFactories/CreateEngagementFileRequestFactory.php
@@ -34,53 +34,17 @@
 </COPYRIGHT>
 */
 
-namespace AidingApp\Engagement\Tests\RequestFactories;
+namespace AidingApp\Engagement\Tests\Tenant\EngagementFile\RequestFactories;
 
-use AidingApp\Contact\Models\Contact;
-use AidingApp\Engagement\Models\EngagementBatch;
-use AidingApp\Notification\Enums\NotificationChannel;
-use App\Models\User;
 use Worksome\RequestFactories\RequestFactory;
 
-class CreateEngagementRequestFactory extends RequestFactory
+class CreateEngagementFileRequestFactory extends RequestFactory
 {
     public function definition(): array
     {
         return [
-            'user' => User::factory()->create(),
-            'recipient' => Contact::factory()->create(),
-            'subject' => fake()->sentence,
-            'body' => ['type' => 'doc', 'content' => [['type' => 'paragraph', 'content' => [['type' => 'text', 'text' => fake()->paragraph]]]]],
-            'scheduledAt' => fake()->dateTimeBetween('-1 year', '-1 day'),
-            'channel' => NotificationChannel::Email,
+            'description' => $this->faker->sentence(3),
+            'file' => $this->faker->file(),
         ];
-    }
-
-    public function deliverNow(): self
-    {
-        return $this->state([
-            'scheduledAt' => null,
-        ]);
-    }
-
-    public function deliverLater(): self
-    {
-        return $this->state([
-            'scheduledAt' => fake()->dateTimeBetween('+1 day', '+1 week'),
-        ]);
-    }
-
-    public function ofBatch(): self
-    {
-        return $this->state([
-            'engagement_batch_id' => EngagementBatch::factory(),
-        ]);
-    }
-
-    public function email(): self
-    {
-        return $this->state([
-            'channel' => NotificationChannel::Email,
-        ]);
     }
 }
